Moved DAG code to its own class
